Separate out functional succcess and failure tests
